+++ components/Product/Product.vue
@@ -2,38 +2,47 @@
   <div class="product" itemscope itemtype="http://schema.org/Product">
     <div class="row">
       <div class="col-12 col-md-6">
-        <h1 itemprop="name">{{ product.name }}</h1>
-        <p itemprop="description">{{ product.description }}</p>
+        <h1 itemprop="name">{{ name }}</h1>
+        <p itemprop="description">{{ description }}</p>
       </div>
       <div class="col-12 col-md-6">
-        <b-img-lazy
-          v-if="product.image"
-          :src="product.image"
+        <b-img
+          v-if="image"
+          :src="cloudinaryify(image)"
+          :srcset="[
+            `${cloudinaryify(image, 110)} 110w`,
+            `${cloudinaryify(image, 220)} 220w`,
+            `${cloudinaryify(image, 540)} 540w`,
+            `${cloudinaryify(image, 1080)} 1080w`,
+          ]"
           class="img-fluid mx-auto d-block"
-          :alt="product.name"
+          :alt="name"
           throttle="100"
+          width="540"
+          height="540"
           itemprop="image"
+          blank-src
+          fluid
+          sizes="(max-width: 768px) 50vw, 100vw"
         />
       </div>
     </div>
     <!-- <div class="row d-block d-md-none">
       <div class="col-12">
-        <h1 itemprop="name">{{ product.name }}</h1>
-        <p itemprop="description">{{ product.description }}</p>
+        <h1 itemprop="name">{{ name }}</h1>
+        <p itemprop="description">{{ description }}</p>
       </div>
     </div> -->
 
-    <share class="py-4" />
-
-    <p v-if="product.recipeYield">
+    <p v-if="recipeYield">
       <strong>Servings:</strong>
-      <span>{{ product.recipeYield }}</span>
+      <span>{{ recipeYield }}</span>
     </p>
-    <div v-if="product.ingredient">
+    <div v-if="ingredient">
       <h2>Ingredients:</h2>
       <ol class="list-group-flush">
         <li
-          v-for="ingredient in product.ingredient"
+          v-for="ingredient in ingredient"
           :key="ingredient"
           class="list-group-item"
         >
@@ -43,14 +52,14 @@
     </div>
 
     <div
-      v-if="product.offers.offers"
+      v-if="offers && offers.offers"
       itemprop="offers"
       itemscope
       itemtype="http://schema.org/AggregateOffer"
     >
       <h2>Sellers</h2>
       <div
-        v-for="(offer, index) in product.offers.offers"
+        v-for="(offer, index) in offers.offers"
         :key="index"
         itemprop="offers"
         itemscope
@@ -73,21 +82,18 @@
 
     <nutrition-fact-table
       v-if="
-        product.additionalProperty &&
-        product.additionalProperty.find(
-          (property) => property.name === 'nutrition',
-        )
+        additionalProperty &&
+        additionalProperty.find((property) => property.name === 'nutrition')
       "
       v-bind="
-        product.additionalProperty.find(
-          (property) => property.name === 'nutrition',
-        )
+        additionalProperty.find((property) => property.name === 'nutrition')
       "
       class="my-4"
     />
 
     <keywords
-      :tags="product.keywords ? product.keywords.split(',') : []"
+      v-if="keywords"
+      :tags="keywords ? keywords.split(',') : []"
       label="Tags"
     />
   </div>
@@ -95,28 +101,59 @@
 
 <script>
 import Keywords from '@/components/Keywords.vue';
-import Share from '@/components/Social/Share';
 import NutritionFactTable from '@/components/Recipe/NutritionFactTable';
 
 export default {
   components: {
     Keywords,
-    Share,
     NutritionFactTable,
   },
   inheritAttrs: false,
+  props: {
+    name: { type: String, required: false },
+    description: { type: String, required: false },
+    suitableForDiet: { type: String, required: false },
+    author: { type: Object, required: false },
+    nutrition: { type: Object, required: false },
+    datePublished: { type: String, required: false },
+    keywords: { type: String, required: false },
+    image: { type: [Array, String], required: false },
+    sameAs: { type: Array, required: false },
+    aggregateRating: {
+      type: Object,
+      default: () => ({
+        '@type': 'AggregateRating',
+        ratingValue: 4.93,
+        ratingCount: 1,
+      }),
+    },
+    offers: { type: Array, default: () => [] },
+    additionalProperty: { type: Array, default: () => [] },
+  },
   computed: {
-    product() {
-      // parse id param to int for id lookup
-      const id = parseInt(this.$route.params.id, 10);
-      return this.$store.getters.getProductById(id);
+    linkData() {
+      return {
+        ...this.$props,
+        '@type': 'Product',
+        // dateModified: this.updatedAt.toISOString(),
+        // updatedAt: undefined,
+      };
+    },
+  },
+  methods: {
+    cloudinaryify(image, width = 540) {
+      if (!image.startsWith('https://res.cloudinary.com')) {
+        return `https://res.cloudinary.com/pocketpasta/image/fetch/w_${width},h_${width},ar_1:1,c_fill,f_auto,q_auto/${image}`;
+      } else {
+        return image;
+      }
     },
   },
   head() {
     return {
       script: [
         {
-          json: this.product,
+          json: this.linkData,
           type: 'application/ld+json',
         },
       ],