+++ layouts/default.vue
@@ -73,7 +73,13 @@
       class="hidden-print-only"
     >
       <v-toolbar-title style="width: 256px;" class="ml-0 pl-3">
-        <v-app-bar-nav-icon aria-label="menu" @click.stop="drawer = !drawer" />
+        <v-app-bar-nav-icon aria-label="menu" @click.stop="drawer = !drawer"
+          ><v-progress-circular
+            v-if="loading"
+            indeterminate
+            size="24"
+            width="2"
+        /></v-app-bar-nav-icon>
         <b-img-lazy
           :src="$icon(32)"
           :srcset="`${$icon(32)} 1x, ${$icon(64)} 2x`"
@@ -91,8 +97,8 @@
     </v-app-bar>
     <v-content>
       <nuxt style="min-height: 100vh;" keep-alive />
-      <the-footer />
     </v-content>
+    <the-footer />
   </v-app>
 </template>
 
@@ -108,6 +114,7 @@ export default {
   },
   data() {
     return {
+      loading: true,
       dialog: false,
       drawer: false,
     };
@@ -162,6 +169,9 @@ export default {
     this.$vuetify.theme.dark = this.$store.getters.getThemeById(
       this.$colorMode.value,
     ).dark;
+    if (process.client) {
+      this.loading = false;
+    }
   },
   head() {
     const i18nSeo = this.$nuxtI18nSeo();