+++ pages/products/_id/index.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="container">
-    <product />
+    <product v-bind="item" />
   </div>
 </template>
 
@@ -11,19 +11,46 @@ export default {
   components: {
     Product,
   },
+  async asyncData(context) {
+    const id = parseInt(context.route.params.id, 10);
+    const item = await context.$content('products', id).fetch();
+    return { id, item };
+  },
   data() {
-    return { id: this.$route.params.id };
+    return { id: this.$route.params.id, item: {} };
   },
   head() {
     return {
+      title: this.item.name,
       link: [
         {
           hid: 'canonical',
           rel: 'canonical',
-          href: `https://pocketpasta.com/products/${this.id}`,
+          href: `${this.baseUrl}/products/${this.id}`,
+        },
+      ],
+      meta: [
+        {
+          hid: 'og:title',
+          name: 'og:title',
+          content: this.item.name,
+        },
+        {
+          hid: 'description',
+          name: 'description',
+          content: this.item.description,
+        },
+        {
+          hid: 'og:description',
+          name: 'og:description',
+          content: this.item.description,
         },
       ],
     };
+  },
+  validate({ params }) {
+    // Must be a number
+    return /^\d+$/.test(params.id);
   },
 };
 </script>