+++ test/pages/recipes.spec.js
@@ -59,7 +59,6 @@ const factory = () =>
       NuxtLink: RouterLinkStub,
     },
     localVue,
-    asyncData: [{ ...recipe, id: 1 }, { id: 2 }, { id: 3 }, { id: 4 }],
   });
 
 describe('recipes', () => {